feat: CytoscapeABE folder renamed to AttractorBifurcationTreeVisualization
